Convert user's value to match with field's type
Before updating configuration file, convert user's value type to
the proper one. As an example, when user enters false from command
line, it produces as a string; however, we want it as boolean, so
convert it to boolean.
Update type.json again. Remove the name field and instead, move it
to the index of the array.
Add a method to LoadJSON to check which fields in the configuration
file has a bad type, and then warn user about them. It's really
useful to prevent from errors while running Dej.
Fix many bugs, especially in the output; and some other code
improvements.