Change the method of Dej
Divide the sniffer into two parts, the reader and the sniffer
itself. The reader reads TCPDump raw packets and save the results
into files. When reading is done, the sniffer goes to read them.

The benefits of this change:
1. User doesn't have to wait for TCPDump reader to end its work
when he wants to stop dej (i.e. runs 'dej stop').
2. Reading raw packets in the background helps to get things done
much faster; because of creating concurrent processes.
3. Future changes will be pretty easier.