Improve the reader and the sniffer
Restart the reader to read from the first file when TCPDump
restarted with a new process. This may be caused by refreshing the
interface, for example.
Improve the sniffer. Instead of sniffing read packets in order,
it will sniff the first read packet by the reader and continue.
Fix 'dej status' and 'dej start' to show 4 processes as "Running!",
not 3 (i.e. the previous state).
Also, other improvements.