Remove RootCommand
Instead of the class, add a protected method to BaseCommand and
every command that must be run as root must call it. The reason of
this change is that the structure of the Symfony Console is not
compatible with the previous structure of Dej commands, and thus
every command must be run as root previously.