Use a custom application instance
Extend from Application class of Symfony, and use it instead of the
default one. The reason is that some functionality should be
customized and it was not possible unless using a custom class.
Improve the output when a command is not defined. Improve
ShellOutput a lot, and use formatters and decorators too. Other
changes, also.