Improve 'dej config'
Add 'dej config list', which creates a table from the options
available to be configured, if terminal is big enough. Also, add a
second method to get the details of an option: 'dej config ?
[option]'.
Add some methods to ShellOutput, such as getting shell height and
width sizes. Other improvements, such as creating an XML file for
configuration options.